+++ src/main/java/net/skinsrestorer/bungee/commands/SkinCommand.java
@@ -37,7 +37,7 @@ import net.skinsrestorer.shared.storage.Config;
 import net.skinsrestorer.shared.storage.CooldownStorage;
 import net.skinsrestorer.shared.storage.Locale;
 import net.skinsrestorer.shared.utils.C;
-import net.skinsrestorer.shared.utils.PlayerWrapper;
+import net.skinsrestorer.api.PlayerWrapper;
 import net.skinsrestorer.shared.utils.SRLogger;
 
 import java.util.concurrent.TimeUnit;
