+++ src/main/java/net/skinsrestorer/bungee/utils/SkinApplierBungee.java
@@ -27,10 +27,10 @@ import net.md_5.bungee.connection.LoginResult;
 import net.md_5.bungee.connection.LoginResult.Property;
 import net.skinsrestorer.bungee.SkinsRestorer;
 import net.skinsrestorer.shared.interfaces.SRApplier;
-import net.skinsrestorer.shared.utils.PlayerWrapper;
+import net.skinsrestorer.api.PlayerWrapper;
 import net.skinsrestorer.shared.utils.ReflectionUtil;
 import net.skinsrestorer.shared.utils.SRLogger;
-import net.skinsrestorer.shared.utils.SkinsRestorerAPI;
+import net.skinsrestorer.api.SkinsRestorerAPI;
 
 import java.io.ByteArrayOutputStream;
 import java.io.DataOutputStream;
