+++ src/main/java/net/skinsrestorer/sponge/utils/SkinApplierSponge.java
@@ -24,9 +24,9 @@ package net.skinsrestorer.sponge.utils;
 import com.flowpowered.math.vector.Vector3d;
 import net.skinsrestorer.shared.exception.SkinRequestException;
 import net.skinsrestorer.shared.interfaces.SRApplier;
-import net.skinsrestorer.shared.utils.PlayerWrapper;
+import net.skinsrestorer.api.PlayerWrapper;
 import net.skinsrestorer.shared.utils.Property;
-import net.skinsrestorer.shared.utils.SkinsRestorerAPI;
+import net.skinsrestorer.api.SkinsRestorerAPI;
 import net.skinsrestorer.sponge.SkinsRestorer;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.data.key.Keys;
