+++ src/main/java/net/skinsrestorer/velocity/command/SkinCommand.java
@@ -35,7 +35,7 @@ import net.skinsrestorer.shared.storage.Config;
 import net.skinsrestorer.shared.storage.CooldownStorage;
 import net.skinsrestorer.shared.storage.Locale;
 import net.skinsrestorer.shared.utils.C;
-import net.skinsrestorer.shared.utils.PlayerWrapper;
+import net.skinsrestorer.api.PlayerWrapper;
 import net.skinsrestorer.velocity.SkinsRestorer;
 
 import java.util.concurrent.TimeUnit;
