+++ src/main/java/net/skinsrestorer/velocity/utils/SkinApplierVelocity.java
@@ -27,9 +27,9 @@ import com.velocitypowered.api.util.GameProfile;
 import com.velocitypowered.api.util.GameProfile.Property;
 import net.skinsrestorer.shared.exception.SkinRequestException;
 import net.skinsrestorer.shared.interfaces.SRApplier;
-import net.skinsrestorer.shared.utils.PlayerWrapper;
+import net.skinsrestorer.api.PlayerWrapper;
 import net.skinsrestorer.shared.utils.SRLogger;
-import net.skinsrestorer.shared.utils.SkinsRestorerAPI;
+import net.skinsrestorer.api.SkinsRestorerAPI;
 import net.skinsrestorer.velocity.SkinsRestorer;
 
 import java.io.ByteArrayOutputStream;
